feat!: upgrade to Kafka 4, removing support for ZooKeeper ACLs
